+++ sequence_member/src/main/java/sequence/sequence_member/global/minio/service/MinioService.java
@@ -31,7 +31,6 @@ import sequence.sequence_member.global.utils.MultipartUtil;
 public class MinioService {
 
     private final FileExtension fileExtension;
-
     private MinioClient minioClient;
 
     @Value("${MINIO_ENDPOINT}")
