+++ sequence_member/src/main/java/sequence/sequence_member/global/utils/FileExtension.java
@@ -12,25 +12,11 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-@Component
-public class FileExtension {
 
-    private String[] extensions = {"png", "jpeg", "jpg", "bmp"};
-    private List<String> allowedFileExtensions = new ArrayList<>(Arrays.asList(extensions));
-
+public interface FileExtension {
     @MethodDescription(description = "파일 확장자 추출")
-    public String getFileExtension(MultipartFile file) {
-        String originalFileName = file.getOriginalFilename();
-        if (StringUtils.hasText(originalFileName)) {
-            return originalFileName.substring(originalFileName.lastIndexOf('.') + 1);
-        }
-        return "";
-    }
+    String getFileExtension(MultipartFile file);
 
     @MethodDescription(description = "파일 확장자를 확인합니다.")
-    public void uploadFileExtensionCheck(String extension) {
-        if (!allowedFileExtensions.contains(extension)) {
-
-        }
-    }
+    void uploadFileExtensionCheck(String extension);
 }
